Flowchart comments
Pytorch objects creation - is the labeling of the arrows really clear?
Maybe it should be explained how does this cross-validation split data
look like? (probably will be when you will be speaking)

Maybe the queue concept should be explained (or is it simply prefetching?). Also, the patched/non patched distinction needs to be made clearly.

What do you think about that? I am saying that from my
perspective of a person who before had only some things to do with both
GANDLF and the data types and problems solved here.